refactor(state): break renderer/mobile import cycle via shared remote/ module
Relocates the desktop-shared snapshot primitives (applyThreadSnapshot,
dispatchRemoteSupervisorEvent, collectRuntimeEventsFromSupervisoryMessage,
shouldReplaceRuntimeItemsFromSnapshot) from @/mobile/storeSync into a new
src/renderer/state/remote/ module, killing the @/mobile -> @/renderer/state ->
@/mobile cycle (grep -rn "@/mobile" src/renderer is now empty). Resolves the
TODO(phase-4) at remoteServersStore.ts:26.

The mobile fan-out (emitTerminalReset/Exited, notifyLiveActivityThreadState,
useGitSummariesStore) is split from the core mutation via RemoteDispatchHooks:
the shared core runs the state mutation then fires optional hooks; mobile wraps
the core with its hooks, desktop calls it with none. Behavior identical on both
paths (mobile side-effect ordering preserved; desktop fan-outs were already
inert). storeSync.ts 316->99. remoteServersStore.test.ts (32) + mobile (193) +
app.test.tsx (14) green.
